+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.0
 
+* Add raw_render() function ton inject yaml without yamlfu rendering
 * Add -e VARNAME1,VARNAME2 to import environment variables (closes #1)
 
 ## 0.0.3